models/AI-Model-Zoo/VAI-1.3-Model-Zoo-Code/TensorFlow/tf_mobilenetv2_cityscapes_1024_2048_132.74G_1.3/requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
gast==0.2.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
importlib-metadata==2.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.2
opencv-python==4.4.0.44
opt-einsum==3.3.0
progressbar==2.5
protobuf==3.13.0
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.2.0
Pillow==8.0.0